+++ stralloc/stralloc_diff.c
@@ -3,7 +3,7 @@
 #include "str.h"
 
 extern int stralloc_diff(const stralloc* a,const stralloc* b) {
-  register unsigned int i;
+  register size_t i;
   register int j;
   for (i=0;;++i) {
     if (i==a->len) return i==b->len?0:-1; if (i==b->len) return 1;
