+++ stralloc/stralloc_ready.c
@@ -8,7 +8,7 @@
  * bytes of space, copies the old string into the new space, frees the
  * old space, and returns 1. Note that this changes sa.s. */
 int stralloc_ready(stralloc *sa,size_t len) {
-  register int wanted=len+(len>>3)+30; /* heuristic from djb */
+  register size_t wanted=len+(len>>3)+30; /* heuristic from djb */
   if (!sa->s || sa->a<len) {
     register char* tmp;
     if (!(tmp=realloc(sa->s,wanted)))
